Added PUGetOprsVarArgParameters in pu-parse-tl.c...
      	This function allows one to pass a variable number of arguments named
after a unique suffix substring of the structure they must
"fill". This needs to be used in conjunction with the macro:
 set_parameter_or_default_value
Example, in your evaluable predicate/function or action you will
now do to parse the terms arguments passed:

learning_actionlib::FibonacciGoal goal;

if (PUGetOprsVarArgParameters(terms, // terms is the arg passed to the action/ep/ef
			      1, //numbers of parameters possibly to fill int the struct
			      set_parameter_or_default_value(goal.order,10,INTEGER)
			      ))
goal.order is the field of a goal structure which will take an
        INTEGER, we will give a default value of 10.

terms in the action call needs to be of the form
(VarArg (order 15)). For example:
(! (Fibonacci (VarArg (order 20))))
order being a suffix of goal.order, we will assign the value 20 to goal.order
(! (Fibonacci (VarArg))) which will then default to 10.
Make sense?

This was made to handle ROS/PR2 Action which takez so many
arguments (while most of them can be defaulted). That way, one can
call these actions with only the arguments which are needed, and
not provide all of them each time. This can of course be used in
others applications.