More cleanup before the release.
Remove reference to malloc.h
Point to some obscur .opf files needed for the doc.
Solve the user-external problem (link on default files).
Everything seems to work, except the relocatable stuff... libtool does not like
it (I was expecting this). Will move to a dl mechanism to load external code in
the kernel.