contrib/src/pu-parse-tl.c contrib/src/pu-parse-tl_f.h: exported some usefull functions which were static before.
src/config.h.in: for some obscure reasons (meaning, I forgot why),
there are a bunch of #undefine at the end of this file which undo what
is done before with the config-ac.h.in (HAVE_SIGACTION was one of them).

src/lex-exp.l.part: remove the "text" version of temporal operator
"?,!,^, etc". hardly used an making a mess in other program namespace
(because the parsing is pretty dumb).

src/oprs-init.c: added a signal handler to clean up properly on SIGINT
(Ctrl-C) before quitting.