Trying to nail down MW bug in slist management.
Maybe he managed to put a NULL Term in the list, confusing the loop
transforming the c_list in lisp_list. Need more investigation, but
this may fix the memory corruption and make the bug more visible.

Add an error check if people are trying to add a NULL element to a
list, as it can create problem when the list is accessed making the
assumption it does not contain any (NULL).