Add option to disable triggered I/O
If setitimer() is available (and it typically is on recent systems) it
will use it to periodically trigger a timer. Only if this timer
triggers, the incoming messages from the server and mp are read and
parsed. The problem is, that the timer interval is not updated if the
user sets main_loop_pool_sec/usec. Also, on recent systems this should
not be necessary as things are speedy enough and due to the use of
select there is no hold-up either. Therefore add an option to disable
the use of setitimer at compile time. If it is enabled, add a little fix
which at least resets the timer if the main loop interval is set.