Fixed a problem with database key structure being over flowed...
The hash keys used in the database had two fields defined as unsigned
char... this was bound to fail one day with somebody asserting a fact
in the DB with thousands of subterm. Not to mention that this must be
highly unefficient... Thanks to Sandra for pointing it and to whoever
wrote the SPARK GenoM module ;-)...

I also clean up the SList data structure to use unsigned int for
length and dynamic. Better safe than sorry.