Adding nth and sprintf functions and associated documentation
'STRING' sprintf (COMPOSED_TERM) will return a STRING which is the
result of the formatted print directives.

'TERM' nth (Integer LISP_LIST) returns the nth TERM of the LISP_LIST,
indexed from zero. An error will occur if the first argument is not an
integer or the list is empty.