Fix relative pathnames in .inc files.
Looks like at some previous time the .inc files were in a parent
directory and the .opf files, etc. were in a data/ subdirectory.  But
now the .inc files are siblings of the .opf files, and the use of
"data/" in relative pathnames is wrong. Removed all the "data/"
prefixes.