Fixed the oprs-mode.el to load in my current emacs (GNU Emacs 23).
It looks like setf is not defined by default. I found it in cl-compat,
but this seems to be obsolete. AFAIK, those setf could be replaced
with setq anyway.

I have not tested the whole thing... but editing .opf, .sym and .inc
seems to provide a nice keyword handling.

Use at your own risks...